Reorganize views for better clarity. Add editing capabilities for Info Items.
