+++ src/teams/templates/info_item_form.html
@@ -1,0 +1,35 @@
+{% extends 'base.html' %}
+{% load commonmark %}
+{% load bootstrap3 %}
+
+{% block title %}
+{% if form.instance.id %}
+Edit Info Item: {{ form.instance.headline }}
+{% else %}
+Create Info item
+{% endif %}
+for in {{ form.instance.category.headline }}
+{% endblock %}
+
+{% block content %}
+<div class="panel panel-default">
+    <div class="panel-heading">
+        <h4>
+            {% if form.instance.id %}
+            Edit Info Item: {{ form.instance.name }}
+            {% else %}
+            Create Info Item
+            {% endif %}
+            for in {{ form.instance.category.headline }}
+        </h4>
+    </div>
+    <div class="panel-body">
+        <form method="POST">
+            {% csrf_token %}
+            {% bootstrap_form form %}
+            <button type="submit" class="btn btn-primary">{% if form.instance.id %}Save{% else %}Create{% endif %}</button>
+        </form>
+    </div>
+    <div class="panel-footer"><i>This info item belongs to the <a href="{% url 'teams:detail' team_slug=team.slug camp_slug=team.camp.slug %}">{{ team.name }} Team</a></i></div>
+</div>
+{% endblock %}
